barcode_reader/requirements.txt
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.2.0
python-barcode==0.11.0
pyzbar==0.1.8
